Week3 Exs done

+++ group12/382266293/src/com/coderising/download/impl/DownloadUtil.java
@@ -5,13 +5,11 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 
 public class DownloadUtil {
-	
-	
+
 	public static long getCurrentTime() {
 		return System.currentTimeMillis();
 	}
-	
-	
+
 	public static void createTempFile(String tempName, int len) {
 		File file = new File(tempName);
 		if (file.exists()) {
@@ -24,7 +22,7 @@ public class DownloadUtil {
 			int length = len;
 			byte[] buffer = new byte[1024];
 			long times = length / 1024;
-			int left = (int) (length % 1024);
+			int left = length % 1024;
 			for (int i = 0; i < times; i++) {
 				temp.write(buffer);
 			}
@@ -42,6 +40,5 @@ public class DownloadUtil {
 			}
 		}
 	}
-	
-	
+
 }
